requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.0.2
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-cors-headers==2.2.0
django-tastypie==0.14.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
python-dateutil==2.7.2
python-mimeparse==1.6.0
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
